# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  goodrich.CreateOrderParams:
    properties:
      comboDeal:
        type: boolean
      date:
        type: string
      idNumber:
        type: string
      items:
        items:
          $ref: '#/definitions/models.GoodrichOrderItem'
        type: array
      leaseID:
        type: string
      notes:
        type: string
      paymentMethod:
        $ref: '#/definitions/models.GoodrichPaymentMethod'
      phoneNumber:
        type: string
      timeSlot:
        description: 'Format: 11:10 am'
        type: string
    type: object
...
  goodrich.UpdateMenuItemParams:
    properties:
      available:
        type: boolean
      quantity:
        type: integer
      quantityLimit:
        type: boolean
    type: object
  goodrich.UpdateOrderParams:
    properties:
      adminNotes:
        type: string
      status:
        $ref: '#/definitions/models.GoodrichOrderStatus'
    type: object
...
  models.AreaOfStudy:
    properties:
      abbreviation:
        type: string
      courses:
        description: Has many courses
        items:
          $ref: '#/definitions/models.Course'
        type: array
      department:
        $ref: '#/definitions/models.Department'
      departmentID:
        description: Belongs to department
        type: integer
      id:
        type: integer
      name:
        type: string
      professors:
        description: Many2Many professors  (computed periodically from courses)
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.BannedUser:
    properties:
      bulletinRead:
        type: boolean
      bulletinWrite:
        type: boolean
      dormtrak:
        type: boolean
      ephcatch:
        type: boolean
      ephmatch:
        type: boolean
      factrak:
        description: If banned from this service. True means user is banned; false
          means user is not banned
        type: boolean
      id:
        type: integer
      reason:
        description: Reason why user was banned
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Belongs to user
      userID:
        type: integer
    type: object
...
  models.Course:
    properties:
      areaOfStudy:
        $ref: '#/definitions/models.AreaOfStudy'
      areaOfStudyID:
        description: Belongs to area of study
        type: integer
      books:
        description: Many2Many books
        items:
          $ref: '#/definitions/models.Book'
        type: array
      factrakScore:
        type: number
      factrakSurveys:
        description: Has many factrak surveys
        items:
          $ref: '#/definitions/models.FactrakSurvey'
        type: array
      id:
        type: integer
      number:
        type: string
      professors:
        description: Has many professors (thru factrak surveys). Ignore this in Gorm.
          Populate this only whenever needed.
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Department:
    properties:
      areasOfStudy:
        description: Has many areas of study
        items:
          $ref: '#/definitions/models.AreaOfStudy'
        type: array
      id:
        type: integer
      name:
        type: string
      users:
        description: Has many professors & staff
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.DormRoom:
    properties:
      area:
        type: integer
      bathroomDesc:
        type: string
      bedAdjustable:
        type: boolean
      closet:
        type: string
      commonRoomAccess:
        type: boolean
      commonRoomDesc:
        type: string
      dorm:
        $ref: '#/definitions/models.Dorm'
      dormID:
        description: Belongs to dorm
        type: integer
      dormtrakReviews:
        description: Has many Dormtrak reviews
        items:
          $ref: '#/definitions/models.DormtrakReview'
        type: array
      faces:
        type: string
      floorNumber:
        type: integer
      flooring:
        type: string
      hc:
        type: boolean
      id:
        type: integer
      keyOrCard:
        type: string
      noise:
        type: string
      number:
        type: string
      picture:
        type: string
      privateBathroom:
        type: boolean
      roomType:
        type: string
      thermostatAccess:
        type: boolean
      users:
        description: Has many students
        items:
          $ref: '#/definitions/models.User'
        type: array
      walkthrough:
        type: boolean
    type: object
...
  models.DormtrakReview:
    properties:
      bathroomDesc:
        type: string
      bedAdjustable:
        type: boolean
      closet:
        type: string
      comment:
        type: string
      commonRoomAccess:
        type: boolean
      commonRoomDesc:
        type: string
      createdTime:
        description: 'Pass the created time: not looked at by GORM'
        type: string
      dormRoom:
        $ref: '#/definitions/models.DormRoom'
      dormRoomID:
        description: Belongs to dorm room
        type: integer
      flooring:
        type: string
      id:
        type: integer
      keyOrCard:
        type: string
      location:
        type: integer
      loudness:
        type: integer
      noise:
        type: string
      privateBathroom:
        type: boolean
      satisfaction:
        type: integer
      thermostatAccess:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Belongs to user (student)
        type: integer
      wifi:
        type: integer
    type: object
  models.Ephcatch:
    properties:
      createdTime:
        description: 'Pass the created time: not looked at by GORM'
        type: string
      id:
        type: integer
      other:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Belongs to other
      otherID:
        type: integer
      seen:
        type: boolean
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Belongs to user
      userID:
        type: integer
    type: object
...
  models.FactrakAgreement:
    properties:
      agrees:
        type: boolean
      factrakSurvey:
        $ref: '#/definitions/models.FactrakSurvey'
      factrakSurveyID:
        description: Belongs to survey
        type: integer
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Belongs to user
        type: integer
    type: object
  models.FactrakSurvey:
    properties:
      agreements:
        description: Has many agreements
        items:
          $ref: '#/definitions/models.FactrakAgreement'
        type: array
      approachability:
        type: integer
      clientAgreement:
        description: |-
          Pass if the client agreed with the survey; not looked at by GORM.
          True means user agreed, false means user disagreed, and null/missing means user does not have any
          agreement/disagreement.
        type: boolean
      comment:
        type: string
      course:
        $ref: '#/definitions/models.Course'
      courseFormat:
        description: Remote, Hybrid, In-Person
        type: string
      courseID:
        description: Belongs to course
        type: integer
      courseStimulating:
        type: integer
      courseWorkload:
        type: integer
      createdTime:
        description: 'Pass the created time: not looked at by GORM'
        type: string
      flagged:
        type: boolean
      gradeReceived:
        type: string
      id:
        type: integer
      leadLecture:
        type: integer
      mentalHealthSupport:
        type: integer
      outsideHelpfulness:
        type: integer
      professor:
        $ref: '#/definitions/models.User'
      professorID:
        description: Belongs to professor
        type: integer
      promoteDiscussion:
        type: integer
      semesterSeason:
        description: Course info data
        type: string
      semesterYear:
        type: integer
      totalAgree:
        description: Not looked at by GORM, just for returning in JSON
        type: integer
      totalDisagree:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Belongs to user (student)
        type: integer
      wouldRecommendCourse:
        type: boolean
      wouldTakeAnother:
        type: boolean
    type: object
...
  models.GoodrichMenuItem:
    properties:
      available:
        description: false if item is out of stock
        type: boolean
      category:
        type: string
      description:
        type: string
      id:
        type: integer
      price:
        type: number
      quantity:
        type: integer
      quantityLimit:
        description: |-
          Number of this menu item left; goes down with every order
          quantity limit describe if there is a limit on quantity
        type: boolean
      title:
        type: string
      type:
        type: string
    type: object
  models.GoodrichOrder:
    properties:
      adminNotes:
        type: string
      comboDeal:
        type: boolean
      date:
        description: 'Format: 2006-01-02'
        type: string
      id:
        type: integer
      idNumber:
        type: string
      items:
        description: this means this field will be ignored in the DB. Use this to
          pull the menu items from ItemList into actual objects in the Model/Controller
          side.
        items:
          $ref: '#/definitions/models.GoodrichOrderItem'
        type: array
      notes:
        type: string
      paymentMethod:
        $ref: '#/definitions/models.GoodrichPaymentMethod'
      phoneNumber:
        type: string
      status:
        $ref: '#/definitions/models.GoodrichOrderStatus'
      timeSlot:
        description: 'Format: 15:04'
        type: string
      totalPrice:
        type: number
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Belongs to user
        type: integer
    type: object
  models.GoodrichOrderItem:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/models.GoodrichMenuItem'
      note:
        type: string
    type: object
  models.GoodrichOrderStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      GoodrichOrderStatusUnknown: Default is unknown
    x-enum-varnames:
    - GoodrichOrderStatusUnknown
    - GoodrichOrderStatusPlaced
    - GoodrichOrderStatusReady
    - GoodrichOrderStatusPaid
  models.GoodrichPaymentMethod:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-comments:
      GoodrichPaymentMethodUnknown: Default is unknown
    x-enum-varnames:
    - GoodrichPaymentMethodUnknown
    - GoodrichPaymentMethodSwipe
    - GoodrichPaymentMethodPoints
    - GoodrichPaymentMethodCreditCard
    - GoodrichPaymentMethodCash
    - GoodrichPaymentMethodSwipePlusCash
    - GoodrichPaymentMethodSwipePlusCreditCard
...
  models.User:
    properties:
      admin:
        type: boolean
      areasOfStudy:
        description: Many2Many Area of Studies iff professor (computed periodically
          from courses)
        items:
          $ref: '#/definitions/models.AreaOfStudy'
        type: array
      atWilliams:
        type: boolean
      bannedUser:
        allOf:
        - $ref: '#/definitions/models.BannedUser'
        description: Has one (or zero) banned users
      campusPhoneEXT:
        description: campus phone extension after 413-597-
        type: string
      campusStatus:
        description: 'Has campus status: either remote or on-campus'
        type: string
      cellPhone:
        type: string
      classYear:
        type: integer
      department:
        $ref: '#/definitions/models.Department'
      departmentID:
        description: Belongs to Department iff professor (fetched from Williams LDAP)
        type: integer
      dormRoom:
        $ref: '#/definitions/models.DormRoom'
      dormRoomID:
        description: Belongs to Dorm Room iff student
        type: integer
      dormVisible:
        type: boolean
      dormtrakReviews:
        description: Has many dormtrak reviews
        items:
          $ref: '#/definitions/models.DormtrakReview'
        type: array
      entry:
        type: string
      ephcatchEligibility:
        type: boolean
      ephcatches:
        description: Has many ephcatches (owner side)
        items:
          $ref: '#/definitions/models.Ephcatch'
        type: array
      ephmatchProfile:
        allOf:
        - $ref: '#/definitions/models.EphmatchProfile'
        description: Has one ephmatch profile
      factrakAdmin:
        type: boolean
      factrakAgreements:
        description: Has many factrak agreements
        items:
          $ref: '#/definitions/models.FactrakAgreement'
        type: array
      factrakScore:
        description: Only for professor retrieval with rankings
        type: number
      factrakSurveyDeficit:
        type: integer
      factrakSurveys:
        description: |-
          As we cannot be both a student and a professor, this combines either a student or a professor's factrak survey.
          We populate this field as a hook AfterFind.
        items:
          $ref: '#/definitions/models.FactrakSurvey'
        type: array
      hasAcceptedDormtrakPolicy:
        type: boolean
      hasAcceptedFactrakPolicy:
        type: boolean
      homeCountry:
        type: string
      homePhone:
        type: string
      homeState:
        type: string
      homeTown:
        type: string
      homeVisible:
        type: boolean
      homeZip:
        type: string
      id:
        type: integer
      major:
        type: string
      name:
        type: string
      nickname:
        type: string
      notificationSettings:
        allOf:
        - $ref: '#/definitions/models.NotificationSettings'
        description: Has one notification settings
      notificationTokens:
        description: Has many notification tokens
        items:
          $ref: '#/definitions/models.NotificationToken'
        type: array
      offCampus:
        description: If user is Off-Campus and thus doesn't have a dorm
        type: boolean
      offCycle:
        type: boolean
      office:
        $ref: '#/definitions/models.Office'
      officeID:
        description: Belongs to Office iff staff/professor
        type: integer
      onCampusSemester:
        description: used to calculate number of factrack surveys needed
        type: integer
      optOutEphcatch:
        type: boolean
      pronoun:
        type: string
      suBox:
        type: string
      tags:
        description: Has many tags
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      type:
        type: string
      unixID:
        type: string
      visible:
        type: boolean
      williamsEmail:
        type: string
      williamsID:
        description: Williams W# ID
        type: string
    type: object
...
  services.BaseErrorResponse:
    properties:
      error:
        $ref: '#/definitions/services.RespError'
      status:
        type: integer
      updateToken:
        type: boolean
    type: object
  services.RespError:
    properties:
      errorCode:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
host: wso.williams.edu
info:
  contact:
    email: wso-dev@wso.williams.edu
    name: WSO Dev
  description: API for WSO services like factrak, facebook, dormtrak, course scheduler,
    and others.
  title: WSO API
  version: 2.0.0
paths:
...
  /factrak/users/{userID}/surveys:
    get:
      consumes:
      - application/json
      deprecated: true
      description: list one user's surveys
      operationId: listUserSurveys
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      - description: Offset Pagination (timestamp)
        in: query
        name: offset
        type: string
      - description: Limit Pagination
        in: query
        name: limit
        type: integer
      - description: Professor ID
        in: query
        name: professorID
        type: integer
      - description: Course ID
        in: query
        name: courseID
        type: integer
      - collectionFormat: csv
        description: Preload (course, professor)
        in: query
        items:
          type: string
        name: preload
        type: array
      - description: Populate Agreement Counts
        in: query
        name: populateAgreements
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FactrakSurvey'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1331":
          description: must be self
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: List user surveys
      tags:
      - factrak
  /goodrich/menu:
    get:
      consumes:
      - application/json
      description: lists all menu items
      operationId: listMenu
      parameters:
      - description: Get all menu items (including unavailable)
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GoodrichMenuItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: List menu
      tags:
      - goodrich
  /goodrich/menu/{itemID}:
    patch:
      consumes:
      - application/json
      description: updates a menu item
      operationId: updateMenuItem
      parameters:
      - description: Update Menu Item Params
        in: body
        name: updateParams
        required: true
        schema:
          $ref: '#/definitions/goodrich.UpdateMenuItemParams'
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GoodrichMenuItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Update menu item
      tags:
      - goodrich
...
  /goodrich/orders:
    get:
      consumes:
      - application/json
      description: lists all orders
      operationId: listOrders
      parameters:
      - description: Offset Pagination
        in: query
        name: offset
        type: integer
      - description: Limit Pagination
        in: query
        name: limit
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Date
        in: query
        name: date
        type: string
      - description: User ID
        in: query
        name: userID
        type: integer
      - collectionFormat: csv
        description: Allowed Status list
        in: query
        items:
          type: string
        name: statuses
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GoodrichOrder'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: List orders
      tags:
      - goodrich
    post:
      consumes:
      - application/json
      description: creates an order
      operationId: createOrder
      parameters:
      - description: Create Order Params
        in: body
        name: createParams
        required: true
        schema:
          $ref: '#/definitions/goodrich.CreateOrderParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GoodrichOrder'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Create order
      tags:
      - goodrich
  /goodrich/orders/{orderID}:
    get:
      consumes:
      - application/json
      description: gets an order
      operationId: getOrder
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GoodrichOrder'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Get order
      tags:
      - goodrich
    patch:
      consumes:
      - application/json
      description: updates an order
      operationId: updateOrder
      parameters:
      - description: Update Order Params
        in: body
        name: updateParams
        required: true
        schema:
          $ref: '#/definitions/goodrich.UpdateOrderParams'
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GoodrichOrder'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Update order
      tags:
      - goodrich
...
  /onboarding/canonical/{unixID}:
    get:
      consumes:
      - application/json
      description: Gets a user from their unix. Onboarding exercise.
      operationId: getOnboardingCanonicalUser
      parameters:
      - description: Unix ID
        in: path
        name: unixID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Gets a user
      tags:
      - onboarding
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"